refacto: separate mutations and queries in specific files (#43)
